Removed autoboxing for 1.4 backwards compat.
git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk@265 ef7698a4-5110-0410-9fc6-c7eb3693863f

+++ wicket-contrib-database/src/java/wicket/contrib/database/DatabaseWebPage.java
@@ -61,7 +61,7 @@ public class DatabaseWebPage extends WebPage
 
 	protected IModel newModel(Class c, Long id)
 	{
-		if (id == -1)
+		if (id.longValue() == -1)
 		{
 			return null;
 		}
@@ -82,7 +82,7 @@ public class DatabaseWebPage extends WebPage
 	
 	protected Object load(Class c, Long id)
 	{
-		if (id == -1)
+		if (id.longValue() == -1)
 		{
 			return null;
 		}
